Refactoring TritonAgent to fix broken deployed code
Rather than passing through all calls on TritonAgent directly to
superagent, TritonAgent will now batch up all the data that has been set
and pass it all through to superagent at once, when the one of the
request-firing methods (asPromise(), then(), end()) are called. The
reason for this change is that superagent actually builds a node http
request internally when things like headers are set. Node parses the URL
and sets the host property for the request, and even if we change the
URL later, the host is still fixed - it doesn't get reparsed. This way
we can write plugins that modify any of these parameters without fear
of those changes not taking effect.

This means that we can no longer just automatically wrap all of
superagent's functions, but it seems necessary in this case.