Handle triton response complete after socket finish
We were approximating this with some timers.  This is more correct.

Reported last-byte time and response size in bytes were previously sometimes
too low due to loss of the race.

The `RequestLocalStorage` cls context doesn't bind automatically to the
socket 'finish' callback, so `RequestLocalStorage` now exports a `bind`
method that's a pass-through to the cls context's bind on the server, and a
noop on the client.