Render elements as their data becomes ready
Right now triton sits waiting for the _next_ element's data to resolve before
rendering _anything_, even if _later_ elements could be rendered _right away_.

This patch renders elements as their data resolves, even if that means going
out of order.  It then _writes_ the elements over the socket in order.

This should let us reduce response time by doing CPU work (rendering) while
we're waiting for responses to come back from data endpoints.  This is work
that we used to postpone unnecessarily.

The tradeoff here is GC churn, since we're storing rendered HTML strings that
we used to just spit over the wire and forget about.  This effect is mitigated
by eliminating references to HTML strings as they're emitted.

If a contiguous block of rendered elements is unlocked by the next element
rendering _after_ those that follow it then the whole block will be written
with only one flush at the end.  This should help with output buffering and
compression windowing.