PLAT-60 Elimate superstitious setTimeout
https://github.com/getify/LABjs/blob/master/LAB.src.js#L115 explains:
   setTimeout() "yielding" prevents some weird race/crash conditions in older browsers

But by yielding, it causes the JS to load very late, after a bunch of other images and stuff. On iOS 9, that seems to tickle a bug that causes the browser to hang for 30s before actually loading the files. When we load the script files earlier, it seems to dodge the bug, and improve performance by a few milliseconds, so let's try it.