PLAT-503: RootElements use `listen` instead of `when`
The `when` property was only sufficient for the first render on the server.
Need an interface that supports updates in the browser once the component is
mounted.

The `listen` property expects a function that:
- Takes a function that will be called on changes.
- Returns an unsubscription function that will be called on unmount.

To support this Stores now provide `when*Emitter` functions that are gated on
the `when*` functions and auto-subscribe to changes once through the gate.

Server render only listens for _first_ emission.  Client-side subsequent
emissions update and force re-render.