Move tests out of core/test
This commit moves tests out of core/test into __tests__ subdirectories
so that tests live closer to the modules that the modules that they
verify. Hopefully this separation will make it easier to differentiate
between specs and their support modules as well.

Page integration tests are moved to core/__tests__, because they don't
correspond 1:1 with a particular module.