The path is at `location.pathname`
`location.path` is undefined.  This is a real bug.

Most browsers just default to the current URL when `undefined` is passed as
the third argument to `history.pushstate`, which masks the bug (makes it
irrelevant, actually).

IE11 replaces the last path element with the string "undefined".

Example: "/navigation/playground" becomes "/navigation/undefined".