Move devDependencies up to top level
This speeds up bootstrap and reduces storage requirements.

These packages are developed in the context of the monorepo.  No need for them
to independently install their dev dependencies.  This also helps us stay on
common versions across packages.

This required updating the top-level eslint config, and fixing some eslint
errors that the old version let slip through the cracks in the main
`react-server` package and `react-server-cli`.

Note that I left `react-server` on babel 5.  That's a whole other can of worms.