Remove Sasha's early 2015 data proposal
This pains me, because it's a great doc with an overview of the problem and
the enviroment and a good explanation for where we went with data through most
of 2015.  But, it has gotten out of date and the _stores_ part of it actually
isn't discussing anything that's (currently) part of `react-server`.

This resolves #121.