Standardize routes file configuration
Previously the routes file took a special path through function calls that
prevented it from receiving values from the normal configuration sieve
(defaults => config file => cli).

When the (redundant) CLI default was removed in #206 this meant that a routes
file _must be specified on the command line_.

This patch standardizes handling of routes file configuration.  It also
changes the name of the `--routes` option to `--routes-file` to more
accurately reflect what it is (and to avoid a name conflict with an existing
`const` in `startServer.js`).  As such this is a breaking change.