Busting favicon cache
Some browsers cache favicons for a long time, and Chrome, at least,
is also caching the absence of a favicon. I've waited a few days to
see if the TTL on the favicon absence caching is relatively short,
but it has still not expired. In order to just make it work, this
commit adds a noop version query string parameter to the favicon to
bust the cache.

Part of #447.