Fix timeout error handling
The main issues addressed here are:

1. When renderMiddleware timesout, it will send down what ever it has
left but it fails to close out the body and html.

2. In ClientController there is a race condition between nodeArrival and
failArrival. nodeArrival will receive a bunch of nodes since before
failArrival is sent the nodes we have left are sent down. If failArrival
runs an resolves `retVal` before the nodes can actually render the
`_previouslyRender` flag will be set. Once that is set and the elements
try to render, because `_previouslyRender` is set at the point
`_cleanup` will be called which will wipe away everything on the page.