Add an option to split bundle load from execution
Experiments at Redfin with the `jsBelowTheFold` option have turned up some
improvement in time to display content at the expense of degradation of time
to make the page _interactive_.  This new option is an attempt to retain the
benefit while regaining some of the lost ground in terms of interactivity.

The _load_ of JS bundles is started immediately, but the _parse_ of the
bundles is delayed until `<TheFold/>` is reached.

Pages can opt into this using the response from `handleRoute()`:

```js
handleRoute() {
  ...
  return {code: 200, splitJsLoad: true};
}
```

There is also a `?_debug_split_js_load=1` option that may be used for testing.