Simplify device type routing
The mobile-detect package is large. We shouldn't send it to the browser.

This change uses mobile-detect on the server only. It associates a device
type with the request and sends only that pre-determined value to the browser.

This eliminates the "mobile" route target, which was previously the union of
"phone" and "tablet". These two must be specified individually going forward
(though they may route to the same page class). It also adds a "desktop" route
target.

The `getMobileDetect()` method on the request context is eliminated, and
replaced with a `getDeviceType()` method that returns a string.

This is a breaking change.